backend/requirements.txt
fastapi==0.110.2
google-generativeai==0.8.5
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
sentence-transformers==4.1.0
pandas==2.2.2
pydantic==2.7.0
pinecone==6.0.2
tqdm==4.67.1


